feat: reorganise header routes
