+++ src/components/layout/Header.tsx
@@ -1,14 +1,10 @@
 'use client';
 
-import { MenuItem } from '@mui/material';
-import { Menu } from '@mui/material';
-import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
 import Headroom from 'react-headroom';
 import { useTranslation } from 'react-i18next';
-import { AiFillCaretDown, AiFillCaretUp } from 'react-icons/ai';
 
 import { useOnKeyDown } from '@/hooks/useOnKeyDown';
 
@@ -23,8 +19,9 @@ import UnstyledLink from '@/components/links/UnstyledLink';
 import { MobileMenu } from '@/components/molecules/MobileMenu/MobileMenu';
 
 export const links = [
-  { href: '/projects', label: 'headerMenu.projects' },
+  { href: '/about', label: 'headerMenu.about' },
   { href: '/cv', label: 'headerMenu.cv' },
+  { href: '/projects', label: 'headerMenu.projects' },
   { href: '/uses', label: 'headerMenu.uses' },
   { href: '/contacts', label: 'headerMenu.contact' },
 ];
@@ -35,15 +32,6 @@ export default function Header() {
   const languageDefs = languages as LanguageDef[];
 
   const [isOpen, setIsOpen] = useState(false);
-
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
 
   const pathname = usePathname();
 
@@ -76,40 +64,6 @@ export default function Header() {
             </UnstyledLink>
 
             <ul className='hidden items-center justify-between space-x-6 text-lg md:flex'>
-              <li>
-                <span
-                  className='animated-underline-2 dark:animated-underline flex items-center rounded-sm font-medium text-slate-950 dark:text-blue-50'
-                  aria-controls={open ? 'basic-menu' : undefined}
-                  aria-haspopup='true'
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                >
-                  {t('headerMenu.about')}
-                  {!open && <AiFillCaretDown className='ms-1.5' />}
-                  {open && <AiFillCaretUp className='ms-1.5' />}
-                </span>
-                <Menu
-                  id='basic-menu'
-                  anchorEl={anchorEl}
-                  open={open}
-                  onClose={handleClose}
-                  MenuListProps={{
-                    'aria-labelledby': 'basic-button',
-                  }}
-                >
-                  <MenuItem onClick={handleClose}>
-                    <Link href='/about/work'>
-                      👔 {t('headerMenu.aboutWork')}
-                    </Link>
-                  </MenuItem>
-                  <MenuItem onClick={handleClose}>
-                    <Link href='/about/free-time'>
-                      🪁 {t('headerMenu.aboutFreeTime')}
-                    </Link>
-                  </MenuItem>
-                </Menu>
-              </li>
-
               {links.map(({ href, label }) => (
                 <li key={`${href}${label}`}>
                   <UnstyledLink
